Tarefa2_2018li1g088

Description

Este módulo define funções comuns da Tarefa 2 do trabalho prático.

Testes

testesT2 :: [(Int, Jogada, Estado)]

Testes unitários da Tarefa 2.

Cada teste é um triplo (identificador do Jogador,Jogada a efetuar,Estado anterior).

Funções principais da Tarefa 2.

jogada

Arguments

:: Int
O identificador do Jogador que efetua a jogada.

-> Jogada
A Jogada a efetuar.

-> Estado
O Estado anterior.

-> Estado
O Estado resultante após o jogador efetuar a jogada.

Efetua uma jogada.

muda_posicao_jogador

Arguments

:: Jogada
A Jogada a fazer

-> Estado
O Estado do jogo, será útil para passar para funções auxiliares

-> Jogador
O Jogador a fazer a jogada

-> Jogador
O Jogador resultante após se efeturar a jogada

Muda a posição de um jogador, se todas as condições de movimentação se verificarem

NB: Chamam várias funções auxiliares de verificação de condições, nomeadamente:

verifica_caminho_livre_mapa

verifica_caminho_livre_de_tanques

verifica_caminho_livre_de_choques

direcaoParaVetor

adiciona_disparo

Arguments

:: Int
Identificador do Jogador que disparoui

-> Arma
Arma Disparada

-> [Disparo]
Lista de Disparo

-> Jogador
Jogador que disparou

-> [Disparo]
Lista de Disparo resultante

Adiciona um disparo, consoante a instrução

NB: Chama a função diminui_municao para reduzir a munição do jogador

Funções auxiliares da Tarefa 2.

Funções auxiliares ao movimento

verifica_caminho_livre_mapa :: PosicaoGrelha -> Direcao -> Mapa -> Bool

Verifica se o mapa está livre numa certa direção

verifica_caminho_livre_de_tanques :: Jogada -> Jogador -> Estado -> Bool

Verifica se na direção que o jogador se pretende mexer não há tanques vivos

compara_coordenadas_tanques :: Direcao -> PosicaoGrelha -> PosicaoGrelha -> Bool

Função de auxílio a verifica_caminho_livre_de_tanques que dadas as coordenadas de dois tanques e a direção em que o tanque se quer mover, verifica se o segundo tanque ocupa as posições que o primeiro quer aceder

verifica_caminho_livre_de_choques :: Jogador -> [Jogador] -> [Disparo] -> Bool

Verifica se o tanque está a ser afetado por um choque

NB: Se for o próprio jogados que disparou o choque a mover-se ele consegue fazê-lo

Funções auxiliares ao disparo

diminui_munição :: Arma -> Jogador -> Jogador

Quando um jogador dispara, diminui as suas munições da arma utilizada