Tarefa3_2018li1g088

Description

Este módulo define funções comuns da Tarefa 3 do trabalho prático.

Introdução

testesT3 :: [Estado]

Nesta Tarefa foi-nos colocado o objetivo de comprimir o Estado o mais possível sob a forma de uma string, sendo, posteriormente, possível descomprimir essa string sem que se perdesse informação.

Nesta tarefa o maior desafio foi o de comprimir o mapa pois diferentes métodos teriam diferentes eficácias para diferentes tipos de mapas

Objetivos

Funções principais da Tarefa 3 (Compressão)

comprime :: Estado -> String

Comprime um Estado para formato textual.

A compressão será dividida em três:

Compressão do Mapa: Será usada a função comprime_mapa

Compressão dos Jogadores: Será usada a função comprime_jogadores

Compressao dos Disparos: Será usada a função comprime_disparos

Funções de compressão do mapa

comprime_mapa :: Estado -> String

Dado um Estado, comprime o seu mapa, utilizando a função ord em que o sexto dígito de um caracter corresponde ao tipo de parede enquanto os restantes correspondem ao número de vezes que esse bloco aparece

Passos da compressão do mapa :

Colocar no primeiro caracter a informação do nº de jogadores no dígito das centenas de milhar, o nº de linhas do mapa nas centenas e o nº de colunas nas unidades

Retirar ao mapa as bordas usando a função mapa_borderless

Transformar o mapa numa lista de peças

Transformar essa lista de peças em caracteres, usando a função constroi_string_pecas

Melhorar essa string que comprimiu as peças, usando a função 'melhora_string_peças'

Exemplos de compressão das peças:

\ 100017 corresponde a 17 Vazia

\ 200111 corresponde a 111 Bloco Indestrutivel

\ 400045 corresponde a 45 Vazias seguidas de um Bloco Indestrutível

Nota: Cada caracter da compressão do mapa terá 6 digitos, isto será importante para depois separar a string que comprime o mapa da string que comprime os jogadores, na qual cada caracter só terá 5 dígitos, isto é feito para poupar caracteres de separação da informação

Funções auxiliares à compressão do mapa

melhora_string_pecas :: [Char] -> [Char]

Visto no mapa ocorrer bastantes vezes o caso de após x blocos vazios ocorrer um único bloco Destrutivel ou Industrutível, registou-se com um número esses casos -- -- * 4xxxxx corresponde a xxxxx peças vazias seguidas de um bloco Indestrutível -- -- * 5xxxxx corresponde a xxxxx peças vazias seguidas de um bloco Destrutível

passa_parede_int :: Peca -> Int

Faz corresponder uma peça a um número que servirá para simbolizar essa peça

constroi_string_pecas :: [Peca] -> [Char]

Irá percorrer uma lista onde se encontram todas as peças do mapa e irá comprimi todas as peças iguais e seguidas num único dígito yxxxxx

y - digito correspondente ao caracter, através da função passa_parede_int

xxxx - número de peças seguidas dessa peça

mapa_borderless :: Mapa -> Mapa

Retira as bordas da matriz para melhor compressão

elimina_primeiro_ultimo_lista :: [a] -> [a]

Dada uma lista devolve essa lista sem o primeiro e o último elemento, servirá como auxiliar ao retirar à função mapa_borderless

Funções de compressão de jogadores

comprime_jogadores :: [Jogador] -> String

A informação de cada jogador será comprimida em dois caracteres:

O primeiro contém a direcaoJogador, assim como lasersJogador e choquesJogador

O segundo contém a posicaoJogador e as vidasJogador

Nota: Para comprimir os três valores num só caracter será usada a função comprime_tres_valores

Nota: O primeiro caracter dos dois que comprimem a informnação de um jogador terá 5 dígitos, o que ajudará na separação da string, posteriormente.

comprime_tres_valores

Arguments

:: Int
Int correspondente à Direcao ou vidasJogador

-> Int
lasersJogador ou as linhas das coordenadas

-> Int
choquesJogador ou as colunas das coordenadas

-> Char
Char que nos seus dígitos contém as infromações dadas

Dados três valores devolve um dígito em que nas dezenas de milhar está contida uma informação, nas centenas outra e nas unidades outra

Funções de compressão de disparos

comprime_disparos :: [Disparo] -> String

Comprime uma lista de disparos em string, o caracter DisparoLaser ou um DisparoCanhao terá 6 digitos enquanto o DisparoChoque terá 3

No caso do Canhao e do Laser

\kxyyzz

k: numero dado pela função passa_jogador_disparo_int que indica qual o jogador que disparou assim como o tipo de disparo

x: número correspondente à direção dado por passa_direcao_int

yyxx: y coordenadas do disparo

No caso do Choque

\xy

x: numero do jogador

y: número de ticks

Funções auxiliares à compressão de disparos

junta_disparo_string :: Disparo -> String

Função que fará a transcrição de um disparo para um caracter

Nota: Usará a função passa_jogador_disparo_int e a função passa_direcao_int

passa_jogador_disparo_int :: Disparo -> Int

Função que dado um jogador e um disparo, visto só haver 2*3=8 possibilidades, comprime-as num só digito , (as centenas de milhar). que será adicionado ao resto da informação para formar um caracter.

passa_direcao_int :: Direcao -> Int

Função que dada uma direção devolve um int correspondente

Funções principais da Tarefa 3 (Descompressão)

descomprime :: String -> Estado

Descomprime uma String para um 'Estado.

A descompressão será dividida em três:

Desompressão do Mapa: Será usada a função desccomprime_mapa

Descompressão dos Jogadores: Será usada a função desccomprime_jogadores

Descompressão dos Disparos: Será usada a função descomprime_disparos

Nota: A String será partida nas 3 string correspondentes a cada componente do estado usando a função parte_string duas vezes. A primeira para separar a string correspondente ao mapa do resto, a segunda vez para separar a string dos jogadores da dos disparos

Funções auxiliares à separação de strings

parte_string :: String -> (String, String)

Função que separa a string do mapa da restante

Como sabemos que todos os caracteres da string que comprime o mapa têm 6 digitos e que o primeiro caracter do jogador terá 5, podemos separar a string ao encontrar um caracter com 5 dígitos, ficando assim com uma string correspondente ào mapa e outra correspondente a toda a infromação restante (jogadores e disparos).

parte_string_jogadores_disparos :: Char -> String -> (String, String)

Função que separa a string dos jogadores da dos disparos

O primeiro caracter da string que corresponde ao estado comprimido contém, no dígito das centenas de milhar, o número de Jogadores existentes no estado.

Sendo assim, e sabendo que a informação de cada jogador se encontra em dois caracteres, basta retirar à string que corresponde aos jogadores e disparos (resultante da função parte_string) x caracteres, em que x corresponde ao dobro dos jogadores do Estado.

A Função devolve assim a string separada em duas, usando o take e o drop, em que a primeira será a string dos jogadores e a segunda a string dos disparos.

Funções de descompressão do mapa

descomprime_mapa :: String -> Mapa

Função que reúne todas as outras funções de descompilação do mapa e que faz a ponte entre todas elas

Nota: O primeiro caracter da string contém informação sobre as dimensões do mapa e o número de jogadores e como tal é ignorado na descompressão das peças.

Passos da descompressão:

Reverter a simplificação dos caracteres, usando a função reverte_simplificacao EX: \ 400007 passa a \ 100007 \ 200001

Converter essa string de caracteres em peças, usando a função converte_string_pecas, tendo, como resultado, uma lista de Peças

Retirar, do primeiro caracter da string, as dimensões do mapa, usando a função string_para_dimensoes

Tendo as dimensões do mapa e tendo uma lista de peças, reconstroi-se o mapa sem bordas, usando a função separa_lista_pecas

Por fim, adicionam-se as bordas ao mapa, usando a função reconstroi_mapa

Funções de auxílio à descompressão do mapas

reconstroi_mapa :: [[Peca]] -> Mapa

Função a que é fornecido o resultado da função separa_lista_pecas, que será o equivalente à matriz sem bordas e adiciona-lhe as bordas, para que a matriz do mapa fique completa

reverte_simplificacao :: [Char] -> String

Reverte a simplificação feita pela função melhora_string_pecas

'\400055' = "\100055\200001"

separa_lista_pecas :: Posicao -> [Peca] -> [[Peca]]

Função que, dadas as dimensões do mapa, parte uma lista de peças em várias listas com as dimensões corretas, devolvendo a matriz do mapa sem as bordas

string_para_dimensoes :: String -> Posicao

Função que retorna a dimensão do mapa, que está codificada nos primeiros 4 dígitos do primeiro caracter

'\205439' = (54,39)

converte_string_pecas :: String -> [Peca]

Função que passará a string para uma lista de peças

Nota: Utiliza a função converte_char_peca para processar cada caracter

converte_char_peca :: Char -> [Peca]

Função que passa um caracter para a lista de peças correspondente

'\200002' = [Bloco Indestrutivel, Bloco Indestrutivel]

passa_int_parede :: Int -> Peca

Função que, quando lhe é dado o 6º digito de um caracter, retorna que peça esse digito significava

Funções de descompressão de jogadores

descomprime_jogadores :: String -> [Jogador]

Função que descomprime a informação relativa aos jogadores

Nota: Em cada um dos dois caracteres que codificam um jogador, há três informações, que a função descomprime_tres_valores irá fornecer a esta função

\20304\20201 = (Jogador (2,1) D 2 3 4 )

descomprime_tres_valores :: Char -> (Int, Int, Int)

Função que de um caracter retirará três informações relativas a um jogador

Irá fazendo isto, passsando o caracter par aum valor numérico, usando a função ord, e ,a partir daí, vai dividindo-o sucessivamente para retirar certas partes do número

Funções de descompressão de disparos

descomprime_disparos :: String -> [Disparo]

Função que verifica se o caracter está a conter informação sobre um laser/canhão ousobre um choque e que age de acordo

Nota: Um caracter que codifique um canhao/laser terá mais de 3 dígitos, oq eu não se verifica para um choque que só contém duas infromações

descomprime_disparo :: Char -> Disparo

Os disparos de canhão têm uma identificação em que o sexto dígito é 4 enquanto um laser tem um número igual a 5, logo a função interpretará o caracter consoante esse número

Nota: Chama a função retorna_info

retorna_info :: Char -> (Int, Direcao, Int, Int)

Dado um caracter, retira informação relativa ao jogador que disparou, a direção e as coordenadas

Assim como na descomprime_tres_valores, também esta função usará as propriedades númericas do caracter após ter sido passado para numero

passa_int_direcao :: Int -> Direcao

Passa um inteiro para a direção que estava a codificar

passa_jogador_disparo :: Int -> Int

Passa um inteiro para o inteiro correspondente ao jogador que está a simbolizar

Nota: Os valores vão de 1 a 8 porque, quando se fez a compressão dos disparos, convencionou-se que os jogadores de 0-3 que disparassem Canhoes ficariam com os números 1-4 e os jogadores que disparassem Lasers ficariam com 5-8

Discussão e conclusão

conclusao :: String

Certamente foi uma tarefa interessante, obrigando o aluno o aluno tanto a escrever código efetivo e de qualidade como a esforçar-se para encontrar soluções originais.

O esforço, assim como a criatividade nos métodos de compressão, culminaram numa taxa de compressão média de 98,90%

Em retrospetiva, julgamos os resultados obtidos bastante satisfatórios.