Tarefa4_2018li1g088

Description

Este módulo define funções comuns da Tarefa 4 do trabalho prático.

Testes

testesT4 :: [Estado]

Testes unitários da Tarefa 4.

Cada teste é um Estado.

Funções principais da Tarefa 4.

tick

Arguments

:: Estado
O Estado anterior.

-> Estado
O Estado após um Tick.

Avança o Estado do jogo um Tick de tempo.

NB: Apenas os Disparos afetam o Estado do jogo com o passar do tempo.

NB: Deve chamar as funções tickChoques, tickCanhoes e tickLasers pela ordem definida.

Funções de processamento dos Lasers

tickLasers :: Estado -> Estado

Avança o Estado do jogo um Tick de tempo, considerando apenas os efeitos dos tiros de Laser disparados.

NB: Chama várias funções:

livreLasers - Verifica se já não existem mais DisparoLaser, o que indica que o processamento dos Lasers terminou

removePecaMapaLaser - Elimina os 'Bloco Destrutivel' atingidos por DisparoLasers

removeVidaJogadoresLaser - Reduz a vida dos Jogadores atingidos por DisparoLaser

avancaLasers - Avança os lasers para a próxima posição que devem ocupar

removeDisparoLaserDisparos - Remove os DisparoCanhao atingidos por DisparoLasers

filtraLaserAcabados - Remove os lasers que já atingiram um 'Bloco Indestruivel'

Funções de verificação do processamento

livreLasers :: [Disparo] -> Bool

Verifica se na lista de disparos já não existe nenhum disparoLaser, indicando que o processamento dos lasers acabou

removePecaMapaLaser :: Mapa -> [Disparo] -> Mapa

Dado um mapa, retira todas os blocos destrutíveis atingidos por lasers num determinado tick

Funções de processamento da lista de disparos

filtraLaserAcabados :: [Disparo] -> Mapa -> [Disparo]

Retira da lista todos os lasers que já atingiram uma parede indestrutível

avancaLasers :: [Disparo] -> [Disparo]

Dada uma lista de Lasers, avança-os para a posição que é suposto tomarem de seguida

verificaCaminhoLaser :: Disparo -> Mapa -> Bool

Verifica se um certo laser tem à sua frente uma parede indestrutível, ou se pode continuar a seguir o sue caminho

removeDisparoLaserDisparos :: [Disparo] -> [Disparo]

Dada uma lista de disparos, retira todos os DisparoCanhao que estão a ser atingidos por DisparoLaser

filtroLaser :: [Disparo] -> Disparo -> Bool

Função auxiliar da removeDisparoLaserDisparos, que verifica, para um determinado DisparoCanhao, se este é atingido por algum DisparoLaser

Funções de alteração de jogadores

removeVidaJogadoresLaser :: [Disparo] -> Jogador -> Jogador

Dada uma lista de disparos, sempre que encontra um laser, se as coordenadas que esse laser afeta atingirem o jogador, reduz-lhe o número de vidas

Funções de processamento de Canhoes

tickCanhoes :: Estado -> Estado

Avança o Estado do jogo um Tick de tempo, considerando apenas os efeitos das balas de Canhao disparadas.

NB: Chama várias funções:

processaColisaoDisparoDisparo - retira os disparos de canhão que colidem uns com os outros

removeDisparoCanhaoMapa - retira os disparos de canhão que colidem com paredes

removeDisparoCanhaoJogador - retira os disparos de canhão que dolidem com jogadores

avancaCanhao - avança os disparos de canhão que sobram

removePecaMapaCanhao - remove os blocos Destrutiveis atingidos por disparos de canhao

processaColisaoDisparoJogadores - reduz a vida dos jogadores atingidos

Funções de alteração de jogadores

processaColisaoDisparoJogadores :: [Disparo] -> [Jogador] -> [Jogador]

Dada uma lista de disparos e uma lista de jogadores, retorna uma lista de jogadores com as vidas atualizadas depois de serem atingidas por disparos

removeVidaJogadoresCanhao :: Disparo -> [Jogador] -> [Jogador]

Função auxiliar à processaColisaoDisparoJogadores que, dado um disparo e uma lista de jogadores, retorna uma lista de jogadores depois de afetada por esse disparo

posicoesAVerificarJogadores

Arguments

:: PosicaoGrelha
Posição do Disparo

-> Direcao
Direção do disparo

-> (PosicaoGrelha, PosicaoGrelha, PosicaoGrelha)
Posições que o disparo vai afetar

Retorna as três posições a verificar quando se quer ver se um jogador é atingido pelo disparo

Funções de alteração do mapa

removePecaMapaCanhao :: Mapa -> [Disparo] -> Mapa

Dado um mapa e uma lista de disparos, devolve o mapa depois de seremr retirados os blocos Destrutiveis atingidos

posicoesAVerificarMapa

Arguments

:: PosicaoGrelha
Posição do Disparo

-> Direcao
Direção do disparo

-> (Posicao, Posicao)
Posições do mapa afetadas pelo disparo

Devolve que posições do mapa de deve verificar ao ver se há bloco destrutiveis

Funções de processamento da lista de disparos

removeDisparoCanhaoMapa :: Mapa -> [Disparo] -> [Disparo]

Dada uma lista de disparos e um mapa, devolve

processaColisaoDisparoDisparo :: [Disparo] -> [Disparo]

Remove disparos de Canhao quando colidem com outros disparos de canhão

removeDisparoCanhaoDisparo :: Disparo -> [Disparo] -> ([Disparo], Int)

Função auxiliar à processaColisaoDisparoDisparo, verifica se um disparo Canhao atinge algum disparo canhao na lista, sem ser ele prório e elimina-o, se for o caso.

posicaoAtras :: PosicaoGrelha -> Direcao -> PosicaoGrelha

Indica qual a coordenada das "costas" de um disparo

avancaCanhao :: [Disparo] -> [Disparo]

Avança os disparos de canhão para as posições que devem tomar

removeDisparoCanhaoJogador :: [Disparo] -> [Jogador] -> [Disparo]

Verifica se algum elemento de uma lista de disparos que seja canhão colide com algum jogador e, se for o caso, retira ele disparo da lista

verificaColisaoDisparoJogador :: [Jogador] -> Disparo -> Bool

Função auxiliar à removeDisparoCanhaoJogador, verifica se numa lista de jogadores, algum é atingido por um canhão

Funções de processamento de Choques

tickChoques :: Estado -> Estado

Avança o Estado do jogo um Tick de tempo, considerando apenas os efeitos dos campos de Choque disparados.