Tarefa6_2018li1g088

Description

Este módulo define funções comuns da Tarefa 6 do trabalho prático.

NOTA: Para que a documentaçâo seja gerada corretamente deve ser usado o comando haddock -h -o doc/html Tarefa6_2018li1g088

Introdução

introT6 :: String

Nesta tarefa foi criado um bot para competir contra os bots desenvolvidos pelos outros alunos, assim como para incorporar na Tarefa5, na opção de jogar contra bots.

Foi criada também uma versão mais fraca do bot para na tarefa5 se poder escolher a dificuldade do bot.

Objetivos

Funções principais da Tarefa 6.

bot

Arguments

:: Int
O identificador do Jogador associado ao robot.

-> Estado
O Estado para o qual o robot deve tomar uma decisão.

-> Maybe Jogada
Uma possível Jogada a efetuar pelo robot.

Ordem de condições e funcionamento da função bot

1ª condiçao: Se existir algum jogador vivo à distância de ser afetado por um DisparoChoque, o bot dispara um choque

2ª condição: Se o bot já tiver disparado um choque e o jogador afetado ainda estiver vivo, chama a função reageChoque que fará o bot atacar o jogador afetado pelo choque.

3ª condição: Se existem blocos destrutíveis/jogadores no caminho para o qual o bot está virado, ou numa das outras três direções alternativas (informação que se obtém através das funções existeAlgoNoCaminho e existeAlgoNoutroCaminho), chama a função reageExisteAlgoCaminho

4ª condição: Caso nenhuma das condições acima se verifique, será chamada a função moveJogadorMaisProximo que fará o bot mover-se em direção ao jogador mais próximo.

Nota: Este será o bot de dificuldade máxima na Tarefa5

existeAlgoNoCaminho

Arguments

:: Int
Contador que é iniciado a 0 mas que se torna 1 se forem encontrados blocos destrutiveis e 2 se forem encontrados jogadores.

-> Jogador
Jogador que servirá como "sonda" para ir verificando o caminho

-> Estado
Estado do qual se irão utilizar partes como argumento par outras funções

-> (Bool, Int)
Par de um boleano e um inteiro, o boleano indica se existe algum jogador no caminho, o inteiro se existe algum bloco Destrutivel (2-Jogador, 1-Bloco, 0-Nada)

Função que verifica se no caminho do bot se encontra algum Jogador/Bloco Destrutivel. Os resultados variam consoante o encontrado:

Se tiver um jogador no caminho retorna (True,2)

Se tiver apenas blocos destrutiveis no caminho retorna (False,1)

Se o caminho acabar num bloco Indestrutivel sem nada antes, retorna (False, 0)

Quando a função encontra um bloco Destrutivel, não acaba logo porque quer avaliar se à frente tem algum jogador, sendo assim mais útil usar o laser

Nota: Se na posição imediatamente à frente do bot não encontra nada, avança a posição do jogador por um bloco para testar os blocos seguintes.

ex: Se um jogador Virado para Baixo na posição (3,3) não tem nenhum bloco/jogador nas posições (4,3) e (4,4), o bot avança para a posição (4,3) para testar as posições (5,3) e (5,4). Este procedimento continua até ser encontrado um jogador ou um bloco.

existeAlgoNoutroCaminho :: Int -> Jogador -> Estado -> (Bool, Direcao, Int)

Esta função é chamada em dois casos

Caso o bot não encontre, na direção para a qual está virado, nenhum bloco destrutível ou jogador

Caso o bot apenas encontre, na direção para a qual está virado, blocos destrutiveis. Como tirar vidas a um jogador fornece mais pontos que destruir blocos, vale a pena verificar se noutras direções existem jogadores.

A funnção verificará, para as outras três direções (ou seja, para as direções que o jogador não está virado), se nessas direções existem blocos destrutíveis ou jogadores. Fará isto chamando a função existeAlgoNoCaminho, três vezes, uma para cada direção.

Caso em alguma das direções encontre um jogador, devolve o triplo (True, direção onde o jogador está,2), caso encontre apenas blocos destrutiveis devolve (False, direção onde os blocos estão,1), caso não encontre nem jogadores nem blocos destrutiveis em nenhuma das 4 direções, devolve __(False,C,0)

Nota; Os resultados desta função serão usados na função reageExisteAlgoCaminho

reageExisteAlgoCaminho :: Jogador -> Estado -> Maybe Jogada

Função encarregue de reagir caso exista algum bloco/jogador numa das 4 direções em relação ao jogador

Para saber, em primeiro lugar, se existe algum bloco/jogador, chama as funções existeAlgoNoCaminho e existeAlgoNoutroCaminho

Caso a função existeAlgoNoCaminho indique que existe um jogador no caminho, o bot dispara

Se a função existeAlgoNoCaminho indica que apenas há blocos destrutíveis, mas a função existeAlgoNoutroCaminho indica que há algum jogador noutra direção, mas na mesma linha/coluna, o jogador vira-se para essa coluna.

Caso a função existeAlgoNoCaminho indique que não há nada e a função existeAlgoNoutroCaminho indique que há blocos, vira-se nessa direção

Caso nenhuma das funções indique a existência de jogadores/blocos, o bot não faz nada

Nota: Ao testar outras direções, usa a função roda, definida na tarefa1 para alterar a direção

Funções secundárias do bot

estaBoqueadoDestrutivel :: Mapa -> Jogador -> Bool

Dado o mapa e o bot, verifica se à frente do bot está um blocoIndestrutivel que lhe está a bloquear o caminho

Nota: Será chamada pela função existeAlgoNoCaminho várias vezes até ser encontrada uma parede indestrutivel (o que faz parar a pesquisa pois não interessa o que está para além da parede indestrtivel) ou um jogador. É chamda também a função blocosVizinhos que devolve quais os blocos à frente do jogador

estaBoqueadoIndestrutivel :: Mapa -> Jogador -> Bool

Dado o mapa e o bot, verifica se à frente do bot está um blocoIndestrutivel que lhe está a bloquear o caminho

Nota: Será chamada pela função existeAlgoNoCaminho várias vezes até ser encontrada uma parede. É chamda também a função blocosVizinhos que devolve quais os blocos à frente do jogador.

existemJogadoresCaminho :: PosicaoGrelha -> Direcao -> [Jogador] -> Bool

Dadas as coordenadas e direção para a qual o bot está virado, indica se à frente do bot existe algum jogador

Usará a função posicoesAVerificarJogadores para saber quais as coordenadas a verificar

Nota: Esta função será chamada várias vezes pela função existeAlgoNoCaminho para verificar se na posição à frente do bot existe um jogador

estaEmDistanciaChoque :: Int -> Jogador -> Estado -> Bool

Verifica se existe algum jogador que está suficientemente perto para ser afetado por um choque

Para o fazer, utiliza a função estaProximo que verifica se há algum jogador vivo a distância de 3 ou menos do nosso bot.

existeChoque

Arguments

:: Int
índice do bot

-> [Disparo]

-> Bool
Boleano que indica se o bot disparou um DisparoChoque

Dado o indíce do bot, verifica se éste tem um DisparoChoque ativo

moveJogadorMaisProximo :: Jogador -> Estado -> Maybe Jogada

Após lhe ser dada as coordenadas do jogador mais próximo ao nosso bot, devolve uma jogada no sentido de se movimentar para mais perto dele.

Chama a função encontraJogadorMainProximo para saber quais as coordenadas do jogador que queremos atacar

Chama também a função estaBloqueadoIndestrutivel. Se o bot estiver virado para uma parede esta função dará o valor True, mas nós queremos apenas agir se o bot não estiver contra uma parede. Por essa razão aplicamos o operador not() à função estaBloqueadoIndestrutivel

encontraJogadorMaisProximo :: Jogador -> [Jogador] -> PosicaoGrelha

Devolve as coordenadas do jogador vivo mais próximo ao nosso jogador

Nota: O segundo argumento para a função auxiliar é iniciado a (500,500) porque assim temos a certeza que haverá sempre algum jogador com um valor de coordenadas menor

reageChoque :: Jogador -> Estado -> Maybe Jogada

Chama função jogadorAfetado para conseguir saber informação acerca do jogador que está a ser atingido pelo choque.

De seguida compara as coordenadas.

Se os jogadores não estiverem "alinhados" (ou seja, o bot não consegue acertar no joga caso dispare), o bot move-se em direção ao jogador.

Se, por outro lado, os jogadores já estiverem alinhados (algo que se verifica usando a função compara_coordenadas_tanques), o jogador dispara.

jogadorAfetado :: Jogador -> [Jogador] -> Jogador

Dado um jogador (o nosso bot), que sabemos que está a usar um DisparoChoque e a lista de jogadores, devolve o jogador que está a ser afetado pelo choque.

Para saber qual o jogador que está a ser afetado, basta ver se as coordenadas do jogador que estamos a testar agora se encontram num intervalo de 3 linhas para cimabaixo ou 3 colunas para a direitaesquerda do nosso bot, pois é essa a àrea de ação do choque.

Ignora as coordenadas do próprio bot e também ignora jogadores que já morreram

Nota: O caso Jogador [] serve para a função não falhar quando o jogador que o bot estava a atacar morre

blocosVizinhos :: Mapa -> PosicaoGrelha -> Direcao -> (Peca, Peca)

Sabendo que o bot está a utilizar um choque, esta função dita qual a ação a fazer para conseguir atingir o jogador que está a ser afetado pelo choque.

Devolve as duas peças que se encontram à frente do tanque. Isto será a base para a verificação de muitas funções, incluindo a estaBoqueadoIndestrutivel e a estaBoqueadoDestrutivel

avancaPosicao :: Direcao -> PosicaoGrelha -> PosicaoGrelha

Dada uma direção e uma coordenada devolve a coordenada seguinte

Bots de outras dificuldades

botFacil

Arguments

:: Int
O identificador do Jogador associado ao robot.

-> Estado
O Estado para o qual o robot deve tomar uma decisão.

-> Maybe Jogada
Uma possível Jogada a efetuar pelo robot.

Versão do bot mas de dificuldade fácil

Foi retirada a utilização de choqes e a função reageExisteAlgoCaminho foi alterada para uma versão mais fraca, a reageExisteAlgoCaminhoFacil

reageExisteAlgoCaminhoBotFacil :: Jogador -> Estado -> Maybe Jogada

Versão mais fraca da função reageExisteAlgoCaminho

Nesta versão os Lasers não são utilizados e o bot não prioritiza atacar os jogadores em relação aos blocos

botMedio

Arguments

:: Int
O identificador do Jogador associado ao robot.

-> Estado
O Estado para o qual o robot deve tomar uma decisão.

-> Maybe Jogada
Uma possível Jogada a efetuar pelo robot.

Versão do bot de média dificuldade

| Embora este bot também não use os DisparoChoque, já utiliza a função reageExisteAlgoCaminho, o que significa que irá utilizar DisparoLaser e prioritizar atingir jogadores em relação a blovos

Discussão e Conclusão

conclusaoT6 :: String

Assim como na Tarefa3 e na Tarefa5, a liberdade de desenvolver o bot ao gosto do aluno foi um grande motor de motivação nesta parte final do trabalho.

Julgamos que o os vários bots desenvolvidos se adaptam à dificuldade que é suposto representarem e que enriquecem a experiência de jogo.

Considerações finais e Agradecimentos.

agradecimentos :: String

Em retrospetiva, consideramos toda a experiência envolvendo a cadeiras neste primeiro semestre extremamente satisfatória

A possibilidade de colocar os conhecimentos adquiridos na cadeira de programação funcional em prática revelou-se extremamente cativente e sentimos que nos fez evoluir como estudantes e futuros engenheiros.

Deixamos os nosss agradecimentos a todos os professores da cadeira de Laboratórios de Informática I, especialmente ao Professor Hugo Pacheco, pela sua dedicação e sempre presente paciência para ajudar os alunos.

Sinceramente,

João Correia

Pedro Bastos